Google News RSS Feed/requirements.txt
RSS Collector Core Dependencies
httpx==0.27.0
feedparser==6.0.10
python-dotenv==1.0.1
apscheduler==3.10.4
retry==0.9.2
pytz==2024.1
schedule==1.2.0
beautifulsoup4==4.12.0
lxml==4.9.3
Development Dependencies
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# RSS Collector Dependencies

# Core RSS and web fetching
feedparser>=6.0.10
httpx>=0.24.0

# Date parsing
python-dateutil>=2.8.2

# Scheduling
APScheduler>=3.10.0
schedule>=1.2.0
pytz>=2023.3

# Configuration and data handling
# (json is built-in to Python)

# Logging and utilities
# (logging, hashlib, datetime, os, sys are built-in)

# Optional: Enhanced HTTP handling
certifi>=2023.0.0
urllib3>=1.26.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
